requirements.txt
APScheduler==3.6.3
asgiref==3.7.2
backports.zoneinfo==0.2.1
cachetools==4.2.2
certifi==2023.7.22
Django==4.2.5
djangorestframework==3.14.0
methodism==0.2.1
mysqlclient==2.2.0
Pillow==10.0.1
python-telegram-bot==13.15
pytz==2023.3.post1
six==1.16.0
sqlparse==0.4.4
tornado==6.1
typing-extensions==4.8.0
tzdata==2023.3
tzlocal==5.0.1
